// src/api/generated/models/requestCitizenResponseAction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Pauses staff processing until the citizen answers a question.
 */
export interface RequestCitizenResponseAction {
  action: 'REQUEST_CITIZEN_RESPONSE'
  /**
   * @minLength 3
   * @maxLength 1000
   */
  question: string
}